requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
certifi==2021.5.30
chardet==4.0.0
click==8.0.0
coverage==4.5.1
coveralls==1.5.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.2
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.11.2
djangorestframework-filters==0.10.2
docopt==0.6.2
flake8==3.5.0
greencache==1.0.0
gunicorn==19.9.0
idna==2.10
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.1
pluggy==0.7.1
psycopg2-binary==2.8.5
pycodestyle==2.3.1
pyflakes==1.6.0
python-decouple==3.4
pytz==2021.1
redlib==1.5.7
regex==2021.4.4
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
